nvme/tcp: Add pattern support
When a pattern of "\xcc\xcc\xcc\xcc\xcc\xcc\xcc\xcc" is encountered,
then we can assume that the data is payload which is ignored  and can
skipped. As a result, if the nvmeotcp_req parameter is set, then we skip
all blocks that start with this pattern.

This commit allows running an ext4 filesystem mounted with zerocopy
support.

Signed-off-by: Boris Pismenny <BorisPismenny@gmail.com>